requirements.txt
beautifulsoup4~=4.9.3
Django==3.1.3
gunicorn==20.0.4
python-dotenv==0.15.0
python-telegram-bot==13.0
requests==2.25.0
sentry-sdk==0.19.3
whitenoise==5.2.0
